Fix some warnings

+++ src/core/buffer.c
@@ -155,7 +155,7 @@ int32_t buffer_read_i32(buffer *buf)
     }
 }
 
-size_t buffer_read_raw(buffer *buf, void *value, int max_size)
+size_t buffer_read_raw(buffer *buf, void *value, size_t max_size)
 {
     size_t size = buf->size - buf->index;
     if (size > max_size) {
@@ -166,7 +166,7 @@ size_t buffer_read_raw(buffer *buf, void *value, int max_size)
     return size;
 }
 
-void buffer_skip(buffer *buf, int size)
+void buffer_skip(buffer *buf, size_t size)
 {
     buf->index += size;
 }
